Fix permissions for access to the Maven home dir
When installing Maven manually, all files will be under a single
directory. When installing Maven using a Linux package manager, some
files will be symlinks pointing outside of the Maven home directory.
This is similar to the situation for the Java home directory, so we need
to use the same approach, i.e. generate permissions allowing access to the
symlink targets.